feat: 引入雪花 ID 生成器并重构数据模型
- 新增雪花 ID 生成器，生成 53-bit 兼容的分布式 ID
- 将所有数据表的 ID 类型从 INT 迁移到 BIGINT
- 新增数据库迁移工具，支持现有数据的无缝迁移
- 新增维护服务 (maintenance service)
- 更新所有 Repository、Service 和 Handler 以适配 int64 类型 ID
- 新增迁移 SQL 脚本 (migrations/20260312_snowflake_ids.sql)

BREAKING CHANGE: 所有 API 接口的 ID 参数类型从 int 改为 int64

+++ services/config_policy_service.go
@@ -55,7 +55,7 @@ func (s *ConfigPolicyService) GetByName(ctx context.Context, name string) (*data
 }
 
 // GetByID 根据 ID 获取配置策略
-func (s *ConfigPolicyService) GetByID(ctx context.Context, id int) (*database.ConfigPolicy, error) {
+func (s *ConfigPolicyService) GetByID(ctx context.Context, id int64) (*database.ConfigPolicy, error) {
 	return s.policyRepo.GetByID(ctx, id)
 }
 
@@ -87,7 +87,7 @@ func (s *ConfigPolicyService) Update(ctx context.Context, policy *database.Confi
 }
 
 // Delete 删除配置策略
-func (s *ConfigPolicyService) Delete(ctx context.Context, id int) error {
+func (s *ConfigPolicyService) Delete(ctx context.Context, id int64) error {
 	return s.policyRepo.Delete(ctx, id)
 }
 
@@ -112,7 +112,7 @@ func (s *ConfigPolicyService) RecordAccess(ctx context.Context, log *database.Co
 }
 
 // ListAccessLogs 获取指定策略最近访问日志
-func (s *ConfigPolicyService) ListAccessLogs(ctx context.Context, policyID int, limit int) ([]*database.ConfigAccessLog, error) {
+func (s *ConfigPolicyService) ListAccessLogs(ctx context.Context, policyID int64, limit int) ([]*database.ConfigAccessLog, error) {
 	if _, err := s.policyRepo.GetByID(ctx, policyID); err != nil {
 		return nil, err
 	}
