+++ services/subscription_scheduler.go
@@ -69,7 +69,7 @@ func (s *SubscriptionScheduler) run(ctx context.Context) {
 		// 从未同步过，或距上次同步已超过 interval
 		if sub.LastFetchedAt == nil || now.Sub(*sub.LastFetchedAt) >= interval {
 			log.Printf("[scheduler] 触发自动同步: %s（间隔 %v）", sub.Name, interval)
-			go func(id int, name string) {
+			go func(id int64, name string) {
 				syncCtx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
 				defer cancel()
 				if _, err := s.syncService.SyncSubscription(syncCtx, id); err != nil {
